Chg MsgPalPleine, le numéro de lot doit être sur 16bits

+++ 2-Communication/protocole.h
@@ -85,7 +85,7 @@ typedef struct
 
 typedef struct
 {
-   UINT8 numLot;
+   UINT16 numLot;
 } MsgPalPleine;
 
 #define FAIL(x) if(x == -1) {\
